dashboard/requirements.txt
# Framework web
Flask

# Driver ufficiale per Apache Cassandra
cassandra-driver